Services/CarWash_backend/requirement.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
python-dotenv
django-jazzmin
pillow
django-mongodb-backend
django-softdelete
django-tenants
twilio
pyyaml
ua-parser
user-agents
drf-yasg